change Sloth module to visualize
